chore: refactor and minor tweaks
